Temporary fix.
Seems not possible to decode a stream since fail with E_INVALIDDATA,
probably because of missing correct size of compressed stream.

This code is not used from CFE, it uses the plain C version in the cfe/cfe/lzma
directory that probably doesn't check for error and update anyway the destination
buffer.